requirements.txt
Django==6.0.3
djangorestframework==3.17.1
asgiref==3.11.1
environ==1.0
gunicorn==23.0.0
numpy==2.4.3
packaging==26.0
psycopg[binary]==3.2.10
python-dotenv==1.2.2
sqlparse==0.5.5
sentence-transformers==3.4.1
